refactor: convert blobstore-fs to bindgen
