+++ crates/providers/Cargo.toml
@@ -44,6 +44,7 @@ toml = { version = "0.8", default-features = false }
 tracing = { version = "0.1", default-features = false }
 tracing-futures = { version = "0.2", default-features = false }
 tracing-opentelemetry = { version = "0.20", default-features = false }
+tracing-subscriber = { version = "0.3", default-features = false }
 warp = { version = "0.3", default-features = false }
 wascap = { version = "*", path = "../wascap" }
 wasmbus-rpc = { version = "0.15", default-features = false } # TODO: Replace by `wasmcloud-provider-sdk`
