+++ crates/providers/blobstore-fs/wit/deps.lock
@@ -1,0 +1,4 @@
+[blobstore]
+path = "../../../../wit/wasmcloud/blobstore"
+sha256 = "00a9de85c290a463186526ef0bfec6ba20fb6b5378cbcbd90b9492b1abd04863"
+sha512 = "628262d9f8027844c78fa56f435179e24d94abb2e0716680f27e118641c5f9a60d483122a65d70d749c0971a5a67f663f0dd9dce090a533c04cac44e330fa5b9"
