+++ crates/providers/blobstore-fs/wit/deps.toml
@@ -1,0 +1,1 @@
+blobstore = "../../../../wit/wasmcloud/blobstore"
